+++ Source/ScreenManager.cs
@@ -201,7 +201,7 @@ namespace Hacknet
             var viewport = GraphicsDevice.Viewport;
             SpriteBatch.Begin();
             SpriteBatch.Draw(blankTexture, new Rectangle(0, 0, viewport.Width, viewport.Height),
-                new Color(0, 0, 0, (byte) alpha));
+                new Color((byte)0, (byte)0, (byte)0, (byte) alpha));
             SpriteBatch.End();
         }
